reservations working

+++ Hackovid/app/src/main/java/com/apptest/nensalparc/TimeFraction.kt
@@ -2,7 +2,7 @@ package com.apptest.nensalparc
 
 
 data class TimeFractionModel (
-    val start: Int? = null,
+    val start: Float? = null,
     val duration: Int? = null,
     val maxCapacity: Int? = null,
     val currentCapacity: Int? = null
